[IR] Share `boundValueAt` override generation across reference lowerings
Both the Native and Wasm callable-reference lowerings emitted an
identical `boundValueAt(index)` override. Extract it into a shared
helper in`AbstractFunctionReferenceLowering`.

KT-87447
KT-87445
